Move register-related functions from alinx_ptp to alinx_arch
